Remove itself as allowed role to assume
I wish AWS told what exactly was wrong in `MalformedPolicyDocument` so
have to guess.

My theory is the the role `website-pod-tester` doesn't like to have itself as a principal.
At the momenet of creation the role doesn't exist, obviously.

The `website-pod-tester` configuration was copy&pasted from the
`service-network-tester` where it was added as a principal later. The
role service-network-tester is used by a GHA worker as well as pytest.

Definitely, the roles should be less convoluted.